# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Educational Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  student: 'student',
} as const;

export interface User {
  id: string;
  email: string;
  displayName: string;
  role: UserRole;
  xpAnnual: number;
  xpWeekly: number;
  streak: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  photoURL?: string | null;
  createdAt: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  admin: 'admin',
  student: 'student',
} as const;

export interface UserInput {
  email: string;
  displayName: string;
  role: UserInputRole;
  /** @nullable */
  photoURL?: string | null;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  admin: 'admin',
  student: 'student',
} as const;

export interface UserUpdate {
  displayName?: string;
  role?: UserUpdateRole;
  xpAnnual?: number;
  xpWeekly?: number;
  streak?: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  photoURL?: string | null;
}

export interface StreakUpdate {
  streak: number;
}

export interface GrammarRule {
  id: string;
  title: string;
  description: string;
  explanation: string;
  questionCount?: number;
  createdAt: string;
}

export interface GrammarRuleInput {
  title: string;
  description: string;
  explanation: string;
}

export interface GrammarRuleUpdate {
  title?: string;
  description?: string;
  explanation?: string;
}

export interface Question {
  id: string;
  ruleId: string;
  questionText: string;
  options: string[];
  correctAnswer: string;
  /** @nullable */
  hint?: string | null;
  createdAt: string;
}

export interface QuestionInput {
  ruleId: string;
  questionText: string;
  options: string[];
  correctAnswer: string;
  hint?: string;
}

export interface QuestionUpdate {
  questionText?: string;
  options?: string[];
  correctAnswer?: string;
  /** @nullable */
  hint?: string | null;
}

export interface BulkQuestionsInput {
  questions: QuestionInput[];
}

export interface BulkResult {
  created: number;
  failed: number;
  errors?: string[];
}

export interface QuizStartInput {
  userId: string;
  ruleId: string;
  questionCount: number;
}

export interface QuizSession {
  sessionId: string;
  userId: string;
  ruleId: string;
  questions: Question[];
  startedAt: string;
}

export interface AnswerInput {
  questionId: string;
  answerText: string;
  userId: string;
}

export interface AnswerResult {
  correct: boolean;
  correctAnswer: string;
  /** @nullable */
  hint?: string | null;
  xpAwarded: number;
}

export interface QuizCompleteInput {
  userId: string;
  totalCorrect: number;
  totalQuestions: number;
}

export interface QuizResult {
  score: number;
  xpEarned: number;
  passed: boolean;
  newStreak?: number;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  displayName: string;
  xp: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  photoURL?: string | null;
  streak?: number;
}

export type LeaderboardType = typeof LeaderboardType[keyof typeof LeaderboardType];


export const LeaderboardType = {
  annual: 'annual',
  weekly: 'weekly',
} as const;

export interface Leaderboard {
  type: LeaderboardType;
  entries: LeaderboardEntry[];
  updatedAt: string;
}

export interface LeaderboardTitle {
  id: string;
  title: string;
  minRank: number;
  maxRank: number;
  /** @nullable */
  description?: string | null;
}

export interface LeaderboardTitleInput {
  title: string;
  minRank: number;
  maxRank: number;
  description?: string;
}

export interface LeaderboardTitleUpdate {
  title?: string;
  minRank?: number;
  maxRank?: number;
  /** @nullable */
  description?: string | null;
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  info: 'info',
  warning: 'warning',
  success: 'success',
  announcement: 'announcement',
} as const;

export interface Notification {
  id: string;
  title: string;
  message: string;
  active: boolean;
  type?: NotificationType;
  createdAt: string;
}

export type NotificationInputType = typeof NotificationInputType[keyof typeof NotificationInputType];


export const NotificationInputType = {
  info: 'info',
  warning: 'warning',
  success: 'success',
  announcement: 'announcement',
} as const;

export interface NotificationInput {
  title: string;
  message: string;
  type?: NotificationInputType;
  active?: boolean;
}

export type NotificationUpdateType = typeof NotificationUpdateType[keyof typeof NotificationUpdateType];


export const NotificationUpdateType = {
  info: 'info',
  warning: 'warning',
  success: 'success',
  announcement: 'announcement',
} as const;

export interface NotificationUpdate {
  title?: string;
  message?: string;
  type?: NotificationUpdateType;
  active?: boolean;
}

export interface AnalyticsOverview {
  totalStudents: number;
  totalQuizzes: number;
  totalQuestions: number;
  averageScore: number;
  activeThisWeek: number;
  topStreak?: number;
}

export interface RuleAnalytics {
  ruleId: string;
  title: string;
  totalAttempts: number;
  correctAttempts: number;
  failureRate: number;
  flagged?: boolean;
}

export interface InfoCard {
  id: string;
  title: string;
  content: string;
  /** @nullable */
  icon?: string | null;
  active: boolean;
  order?: number;
  createdAt: string;
}

export interface InfoCardInput {
  title: string;
  content: string;
  icon?: string;
  active?: boolean;
  order?: number;
}

export interface InfoCardUpdate {
  title?: string;
  content?: string;
  /** @nullable */
  icon?: string | null;
  active?: boolean;
  order?: number;
}

export type ListQuestionsParams = {
ruleId?: string;
};