# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Educational Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /grammar-rules:
    get:
      operationId: listGrammarRules
      tags: [grammar-rules]
      summary: List grammar rules
      responses:
        "200":
          description: List of grammar rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GrammarRule"
    post:
      operationId: createGrammarRule
      tags: [grammar-rules]
      summary: Create grammar rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrammarRuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrammarRule"
  /grammar-rules/{ruleId}:
    get:
      operationId: getGrammarRule
      tags: [grammar-rules]
      summary: Get grammar rule
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Grammar rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrammarRule"
    patch:
      operationId: updateGrammarRule
      tags: [grammar-rules]
      summary: Update grammar rule
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrammarRuleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrammarRule"
    delete:
      operationId: deleteGrammarRule
      tags: [grammar-rules]
      summary: Delete grammar rule
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions (optionally filter by rule)
      parameters:
        - name: ruleId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/bulk:
    post:
      operationId: bulkCreateQuestions
      tags: [questions]
      summary: Bulk create questions from upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkQuestionsInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkResult"
  /questions/{questionId}:
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update question
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete question
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /leaderboard/titles:
    get:
      operationId: listLeaderboardTitles
      tags: [leaderboard]
      summary: List admin-defined leaderboard titles
      responses:
        "200":
          description: Titles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardTitle"
    post:
      operationId: createLeaderboardTitle
      tags: [leaderboard]
      summary: Create leaderboard title
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaderboardTitleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardTitle"
  /leaderboard/titles/{titleId}:
    patch:
      operationId: updateLeaderboardTitle
      tags: [leaderboard]
      summary: Update leaderboard title
      parameters:
        - name: titleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaderboardTitleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardTitle"
    delete:
      operationId: deleteLeaderboardTitle
      tags: [leaderboard]
      summary: Delete leaderboard title
      parameters:
        - name: titleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: createNotification
      tags: [notifications]
      summary: Create notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{notificationId}:
    patch:
      operationId: updateNotification
      tags: [notifications]
      summary: Update notification
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete notification
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /info-cards:
    get:
      operationId: listInfoCards
      tags: [info-cards]
      summary: List info cards
      responses:
        "200":
          description: Info cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InfoCard"
    post:
      operationId: createInfoCard
      tags: [info-cards]
      summary: Create info card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InfoCardInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfoCard"
  /info-cards/{cardId}:
    patch:
      operationId: updateInfoCard
      tags: [info-cards]
      summary: Update info card
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InfoCardUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfoCard"
    delete:
      operationId: deleteInfoCard
      tags: [info-cards]
      summary: Delete info card
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    GrammarRule:
      type: object
      required: [id, title, description, explanation, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        explanation:
          type: string
        questionCount:
          type: integer
        createdAt:
          type: string
    GrammarRuleInput:
      type: object
      required: [title, description, explanation]
      properties:
        title:
          type: string
        description:
          type: string
        explanation:
          type: string
    GrammarRuleUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        explanation:
          type: string
    Question:
      type: object
      required: [id, ruleId, questionText, options, correctAnswer, createdAt]
      properties:
        id:
          type: string
        ruleId:
          type: string
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        hint:
          type: ["string", "null"]
        createdAt:
          type: string
    QuestionInput:
      type: object
      required: [ruleId, questionText, options, correctAnswer]
      properties:
        ruleId:
          type: string
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        hint:
          type: string
    QuestionUpdate:
      type: object
      properties:
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        hint:
          type: ["string", "null"]
    BulkQuestionsInput:
      type: object
      required: [questions]
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuestionInput"
    BulkResult:
      type: object
      required: [created, failed]
      properties:
        created:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: string
...
    LeaderboardTitle:
      type: object
      required: [id, title, minRank, maxRank]
      properties:
        id:
          type: string
        title:
          type: string
        minRank:
          type: integer
        maxRank:
          type: integer
        description:
          type: ["string", "null"]
    LeaderboardTitleInput:
      type: object
      required: [title, minRank, maxRank]
      properties:
        title:
          type: string
        minRank:
          type: integer
        maxRank:
          type: integer
        description:
          type: string
    LeaderboardTitleUpdate:
      type: object
      properties:
        title:
          type: string
        minRank:
          type: integer
        maxRank:
          type: integer
        description:
          type: ["string", "null"]
    Notification:
      type: object
      required: [id, title, message, active, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        message:
          type: string
        active:
          type: boolean
        type:
          type: string
          enum: [info, warning, success, announcement]
        createdAt:
          type: string
    NotificationInput:
      type: object
      required: [title, message]
      properties:
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, warning, success, announcement]
        active:
          type: boolean
    NotificationUpdate:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, warning, success, announcement]
        active:
          type: boolean
...
    InfoCard:
      type: object
      required: [id, title, content, active, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        icon:
          type: ["string", "null"]
        active:
          type: boolean
        order:
          type: integer
        createdAt:
          type: string
    InfoCardInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        icon:
          type: string
        active:
          type: boolean
        order:
          type: integer
    InfoCardUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        icon:
          type: ["string", "null"]
        active:
          type: boolean
        order:
          type: integer