# lib/api-zod/src/generated/types/leaderboardTitle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Educational Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardTitle {
  id: string;
  title: string;
  minRank: number;
  maxRank: number;
  /** @nullable */
  description?: string | null;
}