# lib/api-zod/src/generated/types/questionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Educational Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface QuestionUpdate {
  questionText?: string;
  options?: string[];
  correctAnswer?: string;
  /** @nullable */
  hint?: string | null;
}